Add task handler separation and response in JSON
